replace all ints with ll for good measure

+++ codeforces/864/D.cpp
@@ -8,22 +8,22 @@
 
 using namespace std;
 
-int n, m;
+ll n, m;
 vector<ll> importances;
-unordered_map<int, unordered_map<int, int>> adj_lists;
-vector<vector<int>> queries;
-unordered_set<int> visited;
-unordered_map<int, ll> node_importance;
-unordered_map<int, int> child_counts;
-vector<int> parents;
+unordered_map<ll, unordered_map<ll, ll>> adj_lists;
+vector<vector<ll>> queries;
+unordered_set<ll> visited;
+unordered_map<ll, ll> node_importance;
+unordered_map<ll, ll> child_counts;
+vector<ll> parents;
 
-pair<ll, ll> importance_dfs(int node) {
+pair<ll, ll> importance_dfs(ll node) {
     if (visited.count(node) > 0) {
         return {0, 0};
     }
     visited.insert(node);
     ll importance = importances[node - 1];
-    int node_count = 1;
+    ll node_count = 1;
     for (auto adj : adj_lists[node]) {
         if (visited.count(adj.first) > 0) continue;
         parents[adj.first] = node;
@@ -37,9 +37,9 @@ pair<ll, ll> importance_dfs(int node) {
     return {importance, node_count};
 }
 
-void rotate(int x) {
+void rotate(ll x) {
     if (adj_lists[x].empty()) return;
-    int max_child = max_element(adj_lists[x].begin(), adj_lists[x].end(),
+    ll max_child = max_element(adj_lists[x].begin(), adj_lists[x].end(),
                                 [&](const auto &a, const auto &b) {
                                     return make_pair(a.second, -a.first) < make_pair(b.second, -b.first);
                                 })
@@ -47,7 +47,7 @@ void rotate(int x) {
     ll max_child_importance = node_importance[max_child];
     node_importance[max_child] += node_importance[x] - max_child_importance;
     node_importance[x] -= max_child_importance;
-    int max_child_nodes = child_counts[max_child];
+    ll max_child_nodes = child_counts[max_child];
     child_counts[max_child] += child_counts[x] - max_child_nodes;
     child_counts[x] -= max_child_nodes;
     adj_lists[max_child][x] = child_counts[x];
@@ -57,23 +57,23 @@ void rotate(int x) {
 int main() {
     cin >> n >> m;
     importances.resize(n);
-    for (int i = 0; i < n; ++i) {
+    for (ll i = 0; i < n; ++i) {
         cin >> importances[i];
     }
 
-    for (int i = 1; i <= n; ++i) {
-        adj_lists[i] = unordered_map<int, int>();
+    for (ll i = 1; i <= n; ++i) {
+        adj_lists[i] = unordered_map<ll, ll>();
     }
 
-    for (int i = 0; i < n - 1; ++i) {
-        int fr, to;
+    for (ll i = 0; i < n - 1; ++i) {
+        ll fr, to;
         cin >> fr >> to;
         adj_lists[fr][to] = 1;
         adj_lists[to][fr] = 1;
     }
 
-    for (int i = 0; i < m; ++i) {
-        int a, b;
+    for (ll i = 0; i < m; ++i) {
+        ll a, b;
         cin >> a >> b;
         queries.push_back({a, b});
     }
@@ -81,12 +81,12 @@ int main() {
     parents.resize(n + 1, -1);
     importance_dfs(1);
 
-    for (int node = 2; node <= n; ++node) {
+    for (ll node = 2; node <= n; ++node) {
         adj_lists[node].erase(parents[node]);
     }
 
     for (const auto &query : queries) {
-        int a = query[0], b = query[1];
+        ll a = query[0], b = query[1];
         if (a == 2) {
             rotate(b);
         } else {
